SRR-toolkit
This script contains bash commands that will take SRR runs from the file SRR_Acc_List.txt and use a SRA command "fastq-dump", to generate fastq files for each run. I now realize that I actually wanted to use "fasterq-dump". I have files generated that used the fasterq-dump but that is not part of this script. So try both and know which one you used when proceeding to the next steps, which will be using STAR-aligner to generate a reference human genome. Then after I will attempt to align the fastq reads generated, to the human reference genome.